ổn định nhé

+++ src/app/pages/service/base.table.service.ts
@@ -2,6 +2,7 @@ import { BaseService, FilterRequest, SearchRequest } from '@/pages/service/base.
 import { inject } from '@angular/core';
 import { filter } from 'rxjs/operators';
 import { ToastService } from './toast.service';
+import { ConfirmationService } from 'primeng/api';
 
 export abstract class BaseTableService<T> {
     protected id: string = 'id';
@@ -9,34 +10,56 @@ export abstract class BaseTableService<T> {
     protected pageNo: number = 1;
     protected pageSize: number = 10;
     toast = inject(ToastService)
+    confirmService = inject(ConfirmationService);
 
     protected constructor(protected service: BaseService<T>) {
     }
 
     onDelete(item: any) {
+        this.confirmService.confirm({
+            message: 'Bạn có chắc chắn muốn xoá bản ghi này?',
+            header: 'Xác nhận xoá',
+            icon: 'pi pi-exclamation-triangle',
+            acceptLabel: 'Có',
+            rejectLabel: 'Không',
 
-        this.service.delete(item.id).subscribe({
-            next: data => {
-                this.toast.success("Xoá dữ liệu thành công");
-                this.filter()
-            },
-            error: err => {
-                this.toast.error(err)
+            accept: () => {
+                this.service.delete(item.id).subscribe({
+                    next: data => {
+                        this.toast.success("Xoá dữ liệu thành công");
+                        this.filter()
+                    },
+                    error: err => {
+                        this.toast.error(err)
+                    }
+                })
             }
-        })
+        });
+
     }
 
     onDeleteList(items: any[]){
         const ids = items.map(item => item.id);
-        this.service.deleteList(ids).subscribe({
-             next: data => {
-                this.toast.success("Xoá dữ liệu thành công");
-                this.filter()
-            },
-            error: err => {
-                this.toast.error(err)
+
+        this.confirmService.confirm({
+            message: `Bạn có chắc chắn muốn xoá ${ids.length} bản ghi này?`,
+            header: 'Xác nhận xoá',
+            icon: 'pi pi-exclamation-triangle',
+            acceptLabel: 'Có',
+            rejectLabel: 'Không',
+
+            accept: () => {
+                this.service.deleteList(ids).subscribe({
+                    next: data => {
+                        this.toast.success("Xoá dữ liệu thành công");
+                        this.filter()
+                    },
+                    error: err => {
+                        this.toast.error(err)
+                    }
+                })
             }
-        })
+        });
     }
 
     onSearch(params: any): void  {
